added home page and made some enrolment page changes
