+++ university-student-service/frontend/student-portal/src/app/app.component.html
@@ -3,7 +3,7 @@
   (changeIsLeftSidebarCollapsed)="changeIsLeftSidebarCollapsed($event)"
 />
 
-<app-home
+<app-master
   [isLeftSidebarCollapsed]="isLeftSidebarCollapsed()"
   [screenWidth]="screenWidth()"
 />
