+++ university-student-service/frontend/student-portal/src/app/components/enrolment-info/enrolment-info.component.html
@@ -1,27 +1,29 @@
 <div class="enrolment-container">
   <h2>Master of Computer and Information Sciences</h2>
-  <table>
-    <thead>
-      <tr>
-        <th *ngFor="let header of tableHeaders">{{ header }}</th>
-      </tr>
-    </thead>
-    <tbody>
-      <tr *ngFor="let enrolment of enrolments">
-        <td>{{ enrolment.period }}</td>
-        <td>{{ enrolment.classCode }}</td>
-        <td>
-          <a [routerLink]="['/course-info']" class="course-link">
-            {{ enrolment.courseName }}
-          </a>
-        </td>
-        <td>{{ enrolment.status }}</td>
-        <td>{{ enrolment.starts }}</td>
-        <td>{{ enrolment.location }}</td>
-        <td>{{ enrolment.level }}</td>
-        <td>{{ enrolment.stream }}</td>
-        <td>{{ enrolment.efts }}</td>
-      </tr>
-    </tbody>
-  </table>
+  <div class="table-wrapper">
+    <table>
+      <thead>
+        <tr>
+          <th *ngFor="let header of tableHeaders">{{ header }}</th>
+        </tr>
+      </thead>
+      <tbody>
+        <tr *ngFor="let enrolment of enrolments">
+          <td>{{ enrolment.period }}</td>
+          <td>{{ enrolment.classCode }}</td>
+          <td>
+            <a [routerLink]="['/course-info']" class="course-link">
+              {{ enrolment.courseName }}
+            </a>
+          </td>
+          <td>{{ enrolment.status }}</td>
+          <td>{{ enrolment.starts }}</td>
+          <td>{{ enrolment.location }}</td>
+          <td>{{ enrolment.level }}</td>
+          <td>{{ enrolment.stream }}</td>
+          <td>{{ enrolment.efts }}</td>
+        </tr>
+      </tbody>
+    </table>
+  </div>
 </div>
