+++ university-student-service/frontend/student-portal/src/app/components/enrolment-info/enrolment-info.component.scss
@@ -1,18 +1,23 @@
 /* White card container */
 .enrolment-container {
-  margin: 1rem auto;
   width: 100%;
-  border-radius: 8px;
+  border-radius: 1rem;
+  padding: 1rem 1.5rem 1.5rem;
+  background-color: #ffffff;
   box-sizing: border-box;
+  box-shadow: 0 4px 12px rgba(0, 0, 0, 0.1);
 }
 
 /* Header */
 .enrolment-container h2 {
   font-size: 22px;
   font-weight: 700;
-  margin-top: 1em;
   margin-bottom: 1rem;
   color: #1a202c;
+}
+
+.table-wrapper {
+  overflow-x: auto;
 }
 
 /* Table styles */
