+++ university-student-service/frontend/student-portal/src/app/components/home/home.component.html
@@ -1,4 +1,161 @@
-<div class="body" [ngClass]="sizeClass()">
-  <router-outlet></router-outlet>
+<div class="home-wrapper">
+  <h1 class="welcome-title">Welcome!</h1>
+
+  <div class="home-columns">
+    <!-- LEFT: My Schedule -->
+    <div class="panel schedule-panel">
+      <div class="panel-top-banner"></div>
+      <div class="panel-inner">
+        <h2 class="panel-title">MY SCHEDULE</h2>
+
+        <!-- Calendar Block -->
+        <div class="calendar-widget">
+          <div class="calendar-header">JUNE 2025</div>
+
+          <div class="calendar-row">
+            <div class="calendar-day">M</div>
+            <div class="calendar-day">T</div>
+            <div class="calendar-day">W</div>
+            <div class="calendar-day">T</div>
+            <div class="calendar-day">F</div>
+            <div class="calendar-day">S</div>
+            <div class="calendar-day">S</div>
+          </div>
+
+          <div class="calendar-row">
+            <div class="calendar-date">28</div>
+            <div class="calendar-date">29</div>
+            <div class="calendar-date">30</div>
+            <div class="calendar-date">1</div>
+            <div class="calendar-date">2</div>
+            <div class="calendar-date">3</div>
+            <div class="calendar-date today">4</div>
+          </div>
+        </div>
+
+        <!-- Next Section -->
+        <div class="next-class-section">
+          <h4 class="next-heading">NEXT – Mon 5 May</h4>
+          <div class="next-class-details">
+            <div class="next-title">
+              COMP828/W101A – Statistical Programming for Data Science
+            </div>
+            <div class="next-time">9:00 AM to 10:00 AM Online</div>
+          </div>
+        </div>
+
+        <div class="coming-up-section">
+          <h3 class="coming-up-title">COMING UP</h3>
+          <div class="coming-up-grid">
+            <!-- Session 1 -->
+            <div class="class-card">
+              <div class="calendar-icon">📅</div>
+              <div>
+                <div class="class-date">Monday 05 June 2025</div>
+                <div class="class-title">
+                  COMP828/W101A – Statistical Programming
+                </div>
+                <div class="class-time">10:00 to 11:00 WF203</div>
+              </div>
+            </div>
+
+            <!-- Session 2 -->
+            <div class="class-card">
+              <div class="calendar-icon">📅</div>
+              <div>
+                <div class="class-date">Monday 05 June 2025</div>
+                <div class="class-title">
+                  COMP809/W101A – Data Mining and Machine Learning
+                </div>
+                <div class="class-time">11:00 to 12:00 WS114</div>
+              </div>
+            </div>
+
+            <!-- Session 3 -->
+            <div class="class-card">
+              <div class="calendar-icon">📅</div>
+              <div>
+                <div class="class-date">Monday 05 June 2025</div>
+                <div class="class-title">COMP838/W150A – Deep Learning</div>
+                <div class="class-time">4:00 to 5:00 WS220</div>
+              </div>
+            </div>
+
+            <!-- Session 4 -->
+            <div class="class-card">
+              <div class="calendar-icon">📅</div>
+              <div>
+                <div class="class-date">Tuesday 06 June 2025</div>
+                <div class="class-title">
+                  COMP809/W101A – Data Mining and Machine Learning
+                </div>
+                <div class="class-time">2:00 to 4:00 WZ602</div>
+              </div>
+            </div>
+          </div>
+        </div>
+      </div>
+    </div>
+
+    <!-- RIGHT: Latest News -->
+    <div class="panel news-panel">
+      <div class="panel-top-banner"></div>
+      <div class="panel-inner">
+        <h2 class="panel-title">LATEST NEWS & EVENTS</h2>
+
+        <div class="news-card-grid">
+          <div class="news-card">
+            <img
+              class="news-card-img"
+              src="/newsitem1.png"
+              alt="Student Group Fair"
+            />
+            <div class="news-card-content">
+              <h4 class="news-card-title">Student Group Fair – Thu 20 June</h4>
+              <p class="news-card-text">
+                Discover student clubs and organisations and browse group
+                activities, meet members and find your community on campus.
+              </p>
+              <div class="news-card-date">📅 Thu, June 20, 2025</div>
+            </div>
+          </div>
+
+          <div class="news-card">
+            <img
+              class="news-card-img"
+              src="/newsitem2.png"
+              alt="Meet and Greet"
+            />
+            <div class="news-card-content">
+              <h4 class="news-card-title">
+                International Student Meet & Greet – Wed 12 June
+              </h4>
+              <p class="news-card-text">
+                Kick off the semester by connecting with fellow students from
+                around the world.
+              </p>
+              <div class="news-card-date">📅 Wed, June 12, 2025</div>
+            </div>
+          </div>
+
+          <div class="news-card">
+            <img
+              class="news-card-img"
+              src="/newsitem3.png"
+              alt="Worldwide Cultural Festival"
+            />
+            <div class="news-card-content">
+              <h4 class="news-card-title">
+                University Worldwide Cultural Festival – Thu 8 June
+              </h4>
+              <p class="news-card-text">
+                A celebration of dance, food and music from around the world.
+              </p>
+              <div class="news-card-date">📅 Thu, June 8, 2025</div>
+            </div>
+          </div>
+        </div>
+      </div>
+    </div>
+  </div>
 </div>
-<p>Home page works!</p>
