+++ university-student-service/frontend/student-portal/src/app/components/master/master.component.html
@@ -1,0 +1,3 @@
+<div class="body" [ngClass]="sizeClass()">
+  <router-outlet></router-outlet>
+</div>
