+++ university-student-service/frontend/student-portal/src/app/components/master/master.component.scss
@@ -1,0 +1,21 @@
+.body {
+  width: calc(100% - 5rem);
+  margin-left: 5rem;
+  z-index: 0;
+  position: relative;
+  top: 0;
+  transition: all 0.5s ease;
+  background: #fff;
+  padding: 5rem 3rem 3rem 3rem;
+  box-sizing: border-box;
+}
+
+.body-trimmed {
+  width: calc(100% - 16.5625rem);
+  margin-left: 16.5625rem;
+}
+
+.body-md-screen {
+  width: calc(100% - 5rem);
+  margin-left: 5rem;
+}
